ARN-356 add a child
