+++ app/addRow/index.njk
@@ -16,11 +16,10 @@
 {% endblock %}
 
 {% block beforeContent %}
-    referrer:{{ thisUrl }}a
     <div role="navigation" aria-label="back">
         {{ govukBackLink({
-            href: "/" + assessmentId + "/questiongroup/" + navigation.previous.url,
-            text: "Back to " + navigation.previous.name
+            href: returnUrl,
+            text: "Back"
         }) }}
     </div>
 {% endblock %}
@@ -45,15 +44,14 @@
 
                 <div class="questiongroup-action-buttons">
                     {{ govukButton({
-                        text: 'Save and continue',
+                        text: submitText,
                         classes: 'govuk-!-margin-bottom-3 govuk-!-margin-right-1'
                     }) }}
 
-
                     {{ govukButton({
-                        text: 'Back to assessment progress',
+                        text: 'Cancel',
                         classes: 'govuk-button--secondary',
-                        href: '/' + assessmentId + '/questiongroup/' + groupId + '/summary'
+                        href: returnUrl
                     }) }}
                 </div>
